refactor: Nested plots into Plots directory
