+++ src/components/Tabs/Plots/Plot3d/Plot3dAppearanceControls.tsx
