+++ src/components/Tabs/Plots/Plot3d/Plot3dTab.tsx
@@ -5,15 +5,15 @@ import { useDebounce } from "usehooks-ts";
 import { PlotTabType, SelectionOutlineColorMode, TabType, type VectorFieldData } from "src/colorizer";
 import { getSharedWorkerPool } from "src/colorizer/workers/SharedWorkerPool";
 import LoadingSpinner from "src/components/LoadingSpinner";
-import Plot3dToolbar from "src/components/Tabs/Plot3d/Plot3dToolbar";
 import PlotsTabToolbar from "src/components/Tabs/Plots/PlotsTabToolbar";
 import type { SharedPlotTabProps } from "src/components/Tabs/Plots/types";
 import { useInteractionListener } from "src/hooks";
 import { useViewerStateStore } from "src/state";
 import { FlexColumn } from "src/styles/utils";
 
-import { make3dConeTrace } from "./plot_3d_utils";
 import Plot3d from "./Plot3d";
+import Plot3dToolbar from "./Plot3dToolbar";
+import { make3dConeTrace } from "./plot_3d_utils";
 
 const MINIMUM_BIN_COUNT = 10;
 const RESUME_PLAYBACK_TIMEOUT_MS = 500;
