+++ src/components/Tabs/Plots/Plot3d/plot_3d_utils.ts
