+++ src/components/Tabs/Plots/ScatterPlot/ScatterPlotTab.tsx
@@ -17,8 +17,8 @@ import SelectionDropdown from "src/components/Dropdowns/SelectionDropdown";
 import type { SelectItem } from "src/components/Dropdowns/types";
 import LoadingSpinner from "src/components/LoadingSpinner";
 import PlotsTabToolbar from "src/components/Tabs/Plots/PlotsTabToolbar";
+import ScatterplotToolbar from "src/components/Tabs/Plots/ScatterPlot/ScatterplotToolbar";
 import type { SharedPlotTabProps } from "src/components/Tabs/Plots/types";
-import ScatterplotToolbar from "src/components/Tabs/ScatterPlot/ScatterplotToolbar";
 import { SCATTERPLOT_SYNC_AXIS_KEY, SHORTCUT_KEYS } from "src/constants";
 import { useDebounce, useDebounceRecord, useIsMouseButtonDownRef } from "src/hooks";
 import { colorizeStateSelector } from "src/state";
