+++ src/components/Tabs/Plots/ScatterPlot/ScatterplotToolbar.tsx
@@ -6,9 +6,9 @@ import { ResetViewIconSVG } from "src/assets";
 import { PlotRangeType } from "src/colorizer/types";
 import IconButton from "src/components/Buttons/IconButton";
 import SelectionDropdown from "src/components/Dropdowns/SelectionDropdown";
-import ToggleAverageLineButton from "src/components/Tabs/ScatterPlot/ToggleAverageLineButton";
-import ToggleContoursButton from "src/components/Tabs/ScatterPlot/ToggleContoursButton";
-import ToggleHistogramButton from "src/components/Tabs/ScatterPlot/ToggleHistogramButton";
+import ToggleAverageLineButton from "src/components/Tabs/Plots/ScatterPlot/ToggleAverageLineButton";
+import ToggleContoursButton from "src/components/Tabs/Plots/ScatterPlot/ToggleContoursButton";
+import ToggleHistogramButton from "src/components/Tabs/Plots/ScatterPlot/ToggleHistogramButton";
 import { useViewerStateStore } from "src/state";
 import { StyledVerticalRule } from "src/styles/components";
 import { FlexRowAlignCenter, VisuallyHidden } from "src/styles/utils";
