+++ src/components/Tabs/Plots/ScatterPlot/ToggleAverageLineButton.tsx
