+++ src/components/Tabs/Plots/ScatterPlot/ToggleHistogramButton.tsx
