+++ src/components/Tabs/Plots/TrackPlot/PlotTab.tsx
@@ -8,8 +8,8 @@ import IconButton from "src/components/Buttons/IconButton";
 import ColorRampDropdown from "src/components/Dropdowns/ColorRampDropdown";
 import LoadingSpinner from "src/components/LoadingSpinner";
 import PlotsTabToolbar from "src/components/Tabs/Plots/PlotsTabToolbar";
+import PlotWrapper from "src/components/Tabs/Plots/TrackPlot/PlotWrapper";
 import type { SharedPlotTabProps } from "src/components/Tabs/Plots/types";
-import PlotWrapper from "src/components/Tabs/TrackPlot/PlotWrapper";
 import { useViewerStateStore } from "src/state";
 import { FlexRowAlignCenter, NoSpinnerContainer } from "src/styles/utils";
