+++ tests/lineage_utils.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { getCoparents } from "src/components/Tabs/Lineage/lineage_utils";
+import { getCoparents } from "src/components/Tabs/Plots/Lineage/lineage_utils";
 
 describe("getCoparents", () => {
   it("returns empty array for empty inputs", () => {
